apps/api/pyproject.toml
[project]
name = "mission-control-api"
version = "0.1.0"
description = "Mission Control API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.0",
  "fastapi>=0.115.0",
  "gunicorn>=22.0.0",
  "pgvector>=0.3.0",
  "pypdf>=5.0.0",
  "pydantic-settings>=2.0.0",
  "python-docx>=1.1.0",
  "python-multipart>=0.0.9",
  "psycopg2-binary>=2.9.0",
  "sqlalchemy>=2.0.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "reportlab>=4.0.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
